requirements.txt
streamlit
yfinance
pandas
openpyxl
praw
google-generativeai
requests
python-dotenv
plotly
